--- vllm_omni/model_executor/models/voxcpm/__init__.py
@@ -1,5 +1,0 @@
-from .voxcpm import VoxCPMForConditionalGeneration
-
-__all__ = [
-    "VoxCPMForConditionalGeneration",
-]
